Refactoring codes
