+++ app.rb
@@ -3,6 +3,7 @@ require_relative './student'
 require_relative './teacher'
 require_relative './book'
 require_relative './rental'
+require_relative './methods/add_person'
 
 class App
   def initialize
@@ -19,19 +20,6 @@ class App
 
   private
 
-  def actions
-    puts # Add a blank line before the actions list
-    {
-      1 => 'List all books',
-      2 => 'List all people',
-      3 => 'Create a person',
-      4 => 'Create a book',
-      5 => 'Create a rental',
-      6 => 'List all rentals for a given person id',
-      7 => 'Exit'
-    }
-  end
-
   def action(action)
     action_list = { 1 => 'list_books', 2 => 'list_people', 3 => 'add_person', 4 => 'add_book', 5 => 'add_rental',
                     6 => 'list_rentals_by_person_id', 7 => 'exit' }
@@ -45,14 +33,6 @@ class App
     run # List actions once the user is done with the current action
   end
 
-  def add_person
-    puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    type = gets.chomp.to_i
-
-    puts 'Invalide input' if type != 1 && type != 2
-    add_student if type == 1
-    add_teacher if type == 2
-  end
 
   def list_people(show_index: false)
     @persons.each_with_index do |person, index|
@@ -60,32 +40,6 @@ class App
       result = "#{index}) " + result if show_index
       puts result
     end
-  end
-
-  def add_student
-    print 'Enter the name of the student: '
-    name = gets.chomp.to_s
-    print 'Enter the age of the student: '
-    age = gets.chomp.to_i
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp.to_s.downcase == 'y'
-    student = Student.new(age, name, parent_permission: parent_permission)
-    @persons << student
-
-    puts "Student with id #{student.id} successfully created"
-  end
-
-  def add_teacher
-    print 'Enter the name of the teacher: '
-    name = gets.chomp.to_s
-    print 'Enter the age of the teacher: '
-    age = gets.chomp.to_i
-    print "And teacher's specialization: "
-    specilization = gets.chomp.to_s
-    teacher = Teacher.new(specilization, age, name)
-    @persons << teacher
-
-    puts "Teacher with id #{teacher.id} successfully created"
   end
 
   def list_books(show_index: false)
